refactor: Move data classes to dedicated packages and update imports
This commit reorganizes the location of several data classes and updates their import statements throughout the project.

**Key Changes:**

-   **`MyAnimeListItem.kt`**:
    -   Moved from `com.branwen.mal.data.repo` to `com.branwen.mal.models.domain`.
    -   Imports updated in:
        -   `ListItem.kt`
        -   `MyListViewModel.kt`
        -   `AnimeListScreenContent.kt`
        -   `AnimeRemoteDataSource.kt`
        -   `AnimeLocalDataSource.kt`
-   **`AnimeListEntity.kt`**:
    -   Moved from `com.branwen.mal.data.repo` (where it was previously defined within `AnimeRepository.kt`) to `com.branwen.mal.models.entity`.
    -   The separate `AnimeListEntity.kt` file was created under the new package.
-   **`AnimeLocalDataSource.kt` and `AnimeRemoteDataSource.kt`**:
    -   Import paths in `AppModule.kt` were updated to reflect their move from `com.branwen.mal.data.repo` to `com.branwen.mal.data.local` and `com.branwen.mal.data.remote` respectively (though the actual file move for these data sources happened in a previous commit, this commit updates the `AppModule` to reflect that).
-   **`LoginScreen.kt`**:
    -   Replaced `android.util.Log` with `timber.log.Timber` for logging.
-   **`AnimeListScreenContent.kt`**:
    -   Reduced `verticalArrangement` spacing in `LazyColumn` from `12.dp` to `8.dp`.
-   **`gradle/libs.versions.toml`**:
    -   Updated Android Gradle Plugin (agp) version from `8.11.1` to `8.12.0`.

This refactoring aims to improve project structure and consistency by grouping data models into more specific packages.
