+++ app/src/main/java/com/branwen/mal/data/local/AnimeLocalDataSource.kt
@@ -2,8 +2,8 @@ package com.branwen.mal.data.local
 
 import com.branwen.mal.data.repo.AnimeDao
 import com.branwen.mal.data.repo.AnimeListEntity
-import com.branwen.mal.data.repo.MyAnimeListItem
 import com.branwen.mal.models.AnimeNode
+import com.branwen.mal.models.domain.MyAnimeListItem
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
