+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.11.1"
+agp = "8.12.0"
 coilCompose = "3.2.0"
 coilNetworkOkhttp = "3.2.0"
 converterGson = "3.0.0"
